Bump project version to 0.47.0 and update changelog
This commit updates the `Directory.Build.props` file to change the project version from `0.46.0` to `0.47.0`. It also updates the `changelog.md` file to document the changes made in version `0.47.0`, which was released on 2024-05-19. The changes include the addition of `ReplicableEntity`, a base class for entities that can be replicated in both directions, and a version bump for various NuGet packages.