Feature/replicable entity removal (#14)
* Updated packages and added new class

Updated various packages across multiple projects, including `FastEndpoints`, `System.IdentityModel.Tokens.Jwt`, `Microsoft.AspNetCore.Authentication.JwtBearer`, `FluentValidation`, `Microsoft.EntityFrameworkCore`, and `Npgsql.EntityFrameworkCore.PostgreSQL`. Also added a new class `ReplicableEntity` to track audit information and origin of the replication system.

* Bump project version to 0.47.0 and update changelog

This commit updates the `Directory.Build.props` file to change the project version from `0.46.0` to `0.47.0`. It also updates the `changelog.md` file to document the changes made in version `0.47.0`, which was released on 2024-05-19. The changes include the addition of `ReplicableEntity`, a base class for entities that can be replicated in both directions, and a version bump for various NuGet packages.

* Restructured namespace, updated project version, and removed ReplicableEntity

The `Admin.Service.Features.DatabaseServers.Entities` namespace was restructured in `ServiceDbContext.cs` and the `using` directive was moved to the beginning of the file. The `TenantEntityTypeConfiguration` is no longer applied in `OnModelCreating` method. The `ReplicableEntity` class was removed from the `Yaver.Db` namespace, which may affect classes inheriting from it. The project version was updated to `0.48.0` and the `changelog.md` file was updated to reflect these changes.

* Restructured namespace, updated project version, and removed ReplicableEntity

The `Admin.Service.Features.DatabaseServers.Entities` namespace was restructured in `ServiceDbContext.cs` and the `using` directive was moved to the beginning of the file. The `TenantEntityTypeConfiguration` is no longer applied in `OnModelCreating` method. The `ReplicableEntity` class was removed from the `Yaver.Db` namespace, which may affect classes inheriting from it. The project version was updated to `0.48.0` and the `changelog.md` file was updated to reflect these changes.